Delete custom hook
