+++ project/src/components/catalog/catalog.tsx
@@ -1,8 +1,9 @@
 import SmallFilmCard, { SmallFilmCardSkeleton } from '../small-film-card';
 import CatalogType from '../../types/catalog-type';
 import GenresList from '../genres-list/genres-list';
-import useCatalog from '../../hooks/useCatalog';
-
+import { useSelector } from 'react-redux';
+import { useState } from 'react';
+import { filteredMovieListSelector } from '../../utils/filtered-movie-selector';
 import style from './catalog.module.scss';
 
 const enum CatalogData {
@@ -15,7 +16,9 @@ function Catalog({
   similar = false,
 }: CatalogType): JSX.Element {
   const initialLimit = similar ? CatalogData.SimilarLimit : CatalogData.DefaultLimit;
-  const [list, limit, setLimit] = useCatalog(initialLimit);
+
+  const filteredMovies = useSelector(filteredMovieListSelector);
+  const [limit, setLimit] = useState<number>(initialLimit);
 
   return (
     <section
@@ -28,8 +31,8 @@ function Catalog({
       {genres && <GenresList />}
 
       <div className={style.list}>
-        {list
-          ? list.map(
+        {filteredMovies
+          ? filteredMovies.map(
             (movie, index) =>
               index < limit && (<SmallFilmCard movie={movie} key={movie.id} />))
           : [...Array(initialLimit).keys()].map((key) => (
@@ -37,7 +40,7 @@ function Catalog({
           ))}
       </div>
 
-      {list && list.length >= limit + 1 && !similar && (
+      {filteredMovies && filteredMovies.length >= limit + 1 && !similar && (
         <button
           className={style.button}
           type='button'
